Undid the change on UnshackledMutexTest
