+++ tests/Basics/UnshackledMutexTest.cpp
@@ -24,12 +24,84 @@
 
 #include <Basics/UnshackledMutex.h>
 
-#include "ThreadTestHelper.h"
-
 #include <array>
 #include <atomic>
+#include <condition_variable>
+#include <functional>
 #include <memory>
+#include <mutex>
 #include <thread>
+#include <vector>
+
+namespace {
+
+/**
+ * @brief The constructor of WorkerThread starts a thread, which immediately
+ * starts waiting on a condition variable. The execute() method takes a
+ * callback, which is passed to the waiting thread and is then executed by it,
+ * while execute() returns.
+ */
+struct WorkerThread : std::enable_shared_from_this<WorkerThread> {
+  WorkerThread() = default;
+
+  void run();
+
+  void execute(std::function<void()> callback);
+
+  void stop();
+
+ private:
+  std::thread _thread;
+  std::mutex _mutex;
+  std::condition_variable _cv;
+  std::function<void()> _callback;
+  bool _stopped = false;
+};
+
+void operator<<(WorkerThread& workerThread, std::function<void()> callback);
+
+void WorkerThread::run() {
+  // This can't be done in the constructor (shared_from_this)
+  _thread = std::thread([self = this->shared_from_this()] {
+    auto guard = std::unique_lock(self->_mutex);
+
+    auto wait = [&] {
+      self->_cv.wait(
+          guard, [&] { return self->_callback != nullptr || self->_stopped; });
+    };
+
+    for (wait(); !self->_stopped; wait()) {
+      self->_callback.operator()();
+      self->_callback = {};
+    }
+  });
+}
+
+void WorkerThread::execute(std::function<void()> callback) {
+  ASSERT_TRUE(_thread.joinable());
+  ASSERT_TRUE(callback);
+  {
+    auto guard = std::unique_lock(_mutex);
+    ASSERT_TRUE(!_stopped);
+    ASSERT_FALSE(_callback);
+    std::swap(_callback, callback);
+  }
+
+  _cv.notify_one();
+}
+
+void WorkerThread::stop() {
+  {
+    auto guard = std::unique_lock(_mutex);
+    _stopped = true;
+  }
+
+  _cv.notify_one();
+}
+
+void operator<<(WorkerThread& workerThread, std::function<void()> callback) {
+  return workerThread.execute(std::move(callback));
+}
 
 // Used as memoizable thread indexes
 enum Thread {
@@ -59,7 +131,7 @@ enum Thread {
   OMEGA = 23,
 };
 
-using namespace arangodb::tests;
+}  // namespace
 
 // Note that this test will probably succeed even for std::mutex or others,
 // unless you run it with TSan.
